fix: remove broken editable editbutton
edittable cannot work with tables inside structsections and hence that
editbutton makes no sense. However the button is shown via js and so we
cannot hide it with css. Also the rendering of the content is very
agnostic of the specific content, so we cannot prevent its creation.
This leaves removing it with javascript.

This fixes #1