Fix bug about sections being rendered into struct table
This fixes a bug where the sections were rendered into a struct
aggregation table that showed content in a wikitext field.

This bug was caused by attaching the instructions every time
PARSER_HANDLER_DONE is called.

We are now attaching the instructions only on the first PARSER_HANDLER_DONE
event. That should usually be the one for the main content.