Use more modern PHP features
This removes some of the more outdated PHP syntax (e.g. `array()`)
and makes use of some more modern syntax. The stated minimum PHP
compatibility is PHP 7.2.
Even though PHP 7.2 is already End-of-Life, it is still what is
installed in the Ubuntu 18.04 LTS, so it seems sensible to still support
it.