fix: don't render section inside of struct table
Even with ~~NOCACHE~~ is not reparsed if it has not changed. But the
content of struct aggregation tables is. So the plugin instructions were
wrongly added there as well.
But apparently the Doku_Handler has a status field that can be used to
differentiate the rendering of a section and of a page.

Also this adds the ID to the instructions to prevent it from rendering
the instructions from an included page.